CDRIVER-4766 cast size_t/ssize_t to long
This fixes compilation failure when %ld format specifier is used on 32-bit
platforms.

Ref: https://bugs.debian.org/1055264

+++ src/libmongoc/src/mongoc/mongoc-socket.c
@@ -1313,7 +1313,7 @@ _mongoc_socket_try_sendv (mongoc_socket_t *sock, /* IN */
 #else
                   0);
 #endif
-   TRACE ("Send %ld out of %ld bytes", ret, iov->iov_len);
+   TRACE ("Send %ld out of %ld bytes", (long)ret, (long)iov->iov_len);
 #endif
